ml/lmstudio_model_enum_test/requirements.txt
requests
psutil
GPUtil
py-cpuinfo
